Finish <Line /> component basic render
